1. get time working

+++ rashome/src/main/java/com/rashome/rashome/base/Sensor.java
@@ -10,5 +10,5 @@ public abstract class Sensor {
     private String model;
     private int id;
     private float frequency;
-    private float time;
+    private long time;
 }
